renamed component and updated imports
